+++ docs/src/reference/core/changelog.rst
@@ -8,6 +8,7 @@ Fixed
 *****
 
 * fix config bug that skipped preprocessing (overrides, templates) of dict parsed from YAML
+* ``KlioWriteToAvro`` has been enabled as an output event type (previously missing).
 
 0.2.1 (2020-12-03)
 ------------------